+++ R/moveVis-internal.R
@@ -164,7 +164,7 @@ out <- function(input, type = 1, ll = NULL, msg = FALSE, sign = "", verbose = ge
 #' 
 #' @importFrom grDevices colorRampPalette
 #' @noRd 
-.split <- function(m.df, tail_length = 0, path_size = 1, tail_size = 1){
+.split <- function(m.df, tail_length = 0, path_size = 1, tail_size = 1, tail_colour = "white", trace_show = F, trace_colour = "grey"){
   
   # m.names <- unique(as.character(m.df$name))
   # dummy <- lapply(m.names, function(mn){
@@ -189,7 +189,7 @@ out <- function(input, type = 1, ll = NULL, msg = FALSE, sign = "", verbose = ge
     y.colours <- sapply(unique(y$id), function(x) y[y$id == x,]$colour, simplify = F)
     y$tail_colour <- unlist(mapply(y.cols = y.colours, y.size = table(y$id), function(y.cols, y.size){
       y.ramps <- lapply(unique(y.cols), function(x){
-        f <- colorRampPalette(c(x, "white"))
+        f <- colorRampPalette(c(x, tail_colour))
         rev(f(y.size+4)[1:y.size])
       })
       
@@ -200,6 +200,18 @@ out <- function(input, type = 1, ll = NULL, msg = FALSE, sign = "", verbose = ge
     
     # compute tail size from id count
     y$tail_size <- unlist(lapply(table(y$id), function(x) seq(tail_size, path_size, length.out = x)))
+    
+    if(all(isTRUE(trace_show) & i > tail_size)){
+      
+      y.trace <- m.df[!is.na(match(m.df$frame,1:(min(i.range)-1))),]
+      y.trace$colour <- y.trace$tail_colour <-  trace_colour
+      y.trace$tail_size <- tail_size
+      
+      # join trace, reorder by frame and group by id
+      y <- rbind(y, y.trace)
+      y <- y[order(y$frame),]
+      y <- y[order(y$id),]
+    }
     #y$colour <- factor(as.character(y$colour), level = unique(as.character(m.df$colour)))
     #y$name <- factor(as.character(y$name), level = unique(as.character(m.df$name)))
     
@@ -219,7 +231,7 @@ out <- function(input, type = 1, ll = NULL, msg = FALSE, sign = "", verbose = ge
 #' spatial plot function
 #' @importFrom ggplot2 geom_path aes_string theme scale_fill_identity scale_y_continuous scale_x_continuous scale_colour_manual theme_bw guides guide_legend coord_sf
 #' @noRd 
-.gg_spatial <- function(m.split, gg.bmap, m.df, m.crs, path_size = 3, path_end = "round", path_join = "round", equidistant = T, 
+.gg_spatial <- function(m.split, gg.bmap, m.df, m.crs, path_size = 3, path_end = "round", path_join = "round", path_alpha = 1, equidistant = T, 
                         path_mitre = 10, path_arrow = NULL, print_plot = T, path_legend = T, path_legend_title = "Names"){
   
   # frame plotting function
@@ -227,13 +239,13 @@ out <- function(input, type = 1, ll = NULL, msg = FALSE, sign = "", verbose = ge
     
     ## base plot
     p <- y + geom_path(data = x, aes_string(x = "x", y = "y", group = "id"), size = x$tail_size, lineend = path_end, linejoin = path_join,
-                       linemitre = path_mitre, arrow = path_arrow, colour = x$tail_colour) +  theme_bw() +
+                       linemitre = path_mitre, arrow = path_arrow, colour = x$tail_colour, alpha = path_alpha) +  theme_bw() +
       scale_y_continuous(expand = c(0,0)) + scale_x_continuous(expand = c(0,0)) + coord_sf(crs = m.crs, datum = m.crs)
     
     ## add legend?
     if(isTRUE(path_legend)){
       l.df <- cbind.data.frame(x = x[1,]$x, y = x[1,]$y, name = levels(m.df$name),
-                                  colour = as.character(m.df$colour[sapply(as.character(unique(m.df$name)), function(x) match(x, m.df$name)[1] )]), stringsAsFactors = F)
+                               colour = as.character(m.df$colour[sapply(as.character(unique(m.df$name)), function(x) match(x, m.df$name)[1] )]), stringsAsFactors = F)
       l.df$name <- factor(l.df$name, levels = l.df$name)
       l.df <- rbind(l.df, l.df)